Renamed ValidateExecutable function name
Also moved it and isELF to be available on all platform
